fix: move services into sync-workflows to fix dependency issues
