+++ packages/sync-workflows/activities/do_process_sync_changes.ts
@@ -1,4 +1,4 @@
-import { SyncService } from 'sync-worker/services/sync_service';
+import { SyncService } from '../services/sync_service';
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export type DoProcessSyncChangesArgs = {};
