+++ packages/sync-workflows/activities/sync_records_to_destination.ts
@@ -9,8 +9,8 @@ import { CRMCommonModelType } from '@supaglue/types/crm';
 import { EngagementCommonModelType } from '@supaglue/types/engagement';
 import { ApplicationFailure, Context } from '@temporalio/activity';
 import { pipeline, Readable, Transform } from 'stream';
-import { ApplicationService } from 'sync-worker/services';
 import { logEvent } from '../lib/analytics';
+import { ApplicationService } from '../services';
 
 export type SyncRecordsToDestinationArgs = {
   syncId: string;
