+++ packages/sync-workflows/activities/update_sync_state.ts
@@ -1,5 +1,5 @@
 import { Sync, SyncState } from '@supaglue/types';
-import { SyncService } from 'sync-worker/services/sync_service';
+import { SyncService } from '../services/sync_service';
 
 export type UpdateSyncStateArgs = {
   syncId: string;
