+++ packages/sync-workflows/package.json
@@ -4,6 +4,7 @@
   "packageManager": "yarn@3.4.1",
   "dependencies": {
     "@supaglue/core": "workspace:*",
+    "@supaglue/db": "workspace:^",
     "@supaglue/types": "workspace:^",
     "@temporalio/activity": "^1.6.0",
     "@temporalio/client": "^1.6.0",
